penambahan controller wallet

+++ services/category_service.go
@@ -6,26 +6,26 @@ import (
 	"errors"
 )
 
-type CategoryService interface{
+type CategoryService interface {
 	CreateDefaultCategories() ([]models.Category, error)
-	GetAllCategories(userRole int) ([]models.Category, error)
+	GetAllCategories(userRole float64) ([]models.Category, error)
 }
 
-type categoryService struct{
+type categoryService struct {
 	repo repository.CategoryRepository
 }
 
-func NewCategoryService(r repository.CategoryRepository) CategoryService{
+func NewCategoryService(r repository.CategoryRepository) CategoryService {
 	return &categoryService{repo: r}
 }
 
-func (s *categoryService) CreateDefaultCategories() ([]models.Category, error){
+func (s *categoryService) CreateDefaultCategories() ([]models.Category, error) {
 	newCategory, err := s.repo.CreateDefaultCategories()
 	return newCategory, err
 }
 
-func (s *categoryService) GetAllCategories(userRole int) ([]models.Category, error){
-	if userRole > 2{
+func (s *categoryService) GetAllCategories(userRole float64) ([]models.Category, error) {
+	if userRole > 2 {
 		return nil, errors.New("forbidden: access is denied")
 	}
